# openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: oauth-openapi project
    url: https://github.com/kenanhancer/oauth-openapi
  description: |
    # OAuth 2.0 + OpenID Connect — RFC-Canonical OpenAPI

    This specification describes the HTTP surface of an OAuth 2.0 / OpenID Connect
    authorization server in a vendor-neutral way. Every endpoint and field is
    cross-referenced to the source RFC or specification section.

    ## Covered specifications

    | Spec | Year | Subject |
    |---|---|---|
    | RFC 6749 | 2012 | OAuth 2.0 Authorization Framework |
    | RFC 6750 | 2012 | Bearer Token Usage |
    | RFC 7009 | 2013 | Token Revocation |
    | RFC 7517 | 2015 | JSON Web Key (JWK / JWKS) |
    | RFC 7519 | 2015 | JSON Web Token (JWT) |
    | RFC 7636 | 2015 | PKCE — Proof Key for Code Exchange |
    | RFC 7662 | 2015 | Token Introspection |
    | RFC 8414 | 2018 | Authorization Server Metadata |
    | RFC 8628 | 2019 | Device Authorization Grant |
    | RFC 8693 | 2020 | Token Exchange |
    | RFC 9068 | 2021 | JWT Profile for OAuth 2.0 Access Tokens |
    | RFC 9126 | 2021 | Pushed Authorization Requests (PAR) |
    | RFC 9396 | 2023 | Rich Authorization Requests (RAR) |
    | RFC 9449 | 2023 | DPoP — Demonstrating Proof of Possession |
    | RFC 9700 | 2025 | OAuth 2.0 Security Best Current Practice |
    | OAuth 2.1 (draft) | ongoing | Consolidated modern best practices |
    | OIDC Core 1.0 | 2014 | OpenID Connect identity layer |
...

    ## Modeling notes

...

...

...

...
  license:
...
servers:
...
paths:
  /authorize:
    get:
      description: |
        Initiates a browser-driven authorization request. The user-agent is
        navigated to this URL; the server authenticates the resource owner and
        obtains consent, then redirects the user-agent back to `redirect_uri`
        with a `code` (success) or an `error`.

        Under OAuth 2.1 only `response_type=code` is permitted (Implicit is
        deprecated). PKCE (`code_challenge`) is mandatory.
      externalDocs:
        description: "RFC 6749 §4.1.1, OIDC Core §3.1.2.1"
        url: https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.1
      operationId: authorize
      parameters:
      - $ref: "#/components/parameters/ResponseType"
      - $ref: "#/components/parameters/ClientId"
      - $ref: "#/components/parameters/RedirectUri"
      - $ref: "#/components/parameters/Scope"
      - $ref: "#/components/parameters/State"
      - $ref: "#/components/parameters/CodeChallenge"
      - $ref: "#/components/parameters/CodeChallengeMethod"
      - $ref: "#/components/parameters/Nonce"
      - $ref: "#/components/parameters/Prompt"
      - $ref: "#/components/parameters/LoginHint"
      - $ref: "#/components/parameters/MaxAge"
      - $ref: "#/components/parameters/AcrValues"
      - $ref: "#/components/parameters/ResponseMode"
      - $ref: "#/components/parameters/RequestUri"
      - $ref: "#/components/parameters/AuthorizationDetails"
      - $ref: "#/components/parameters/Resource"
      - $ref: "#/components/parameters/Display"
      - $ref: "#/components/parameters/UILocales"
      - $ref: "#/components/parameters/ClaimsLocales"
      - $ref: "#/components/parameters/IDTokenHint"
      - $ref: "#/components/parameters/Claims"
      - $ref: "#/components/parameters/Request"
      - $ref: "#/components/parameters/DPoPJKT"
      responses:
        "302":
          description: |
            User-agent redirect back to `redirect_uri`. The `Location`
            header carries:
            - on success: `code`, `state`, plus `iss` (RFC 9207) when the AS
              advertises `authorization_response_iss_parameter_supported`.
              For OIDC Hybrid (`response_type=code id_token`) it also carries
              `id_token` in the URL fragment.
            - on failure: `error`, `error_description`, `error_uri`, `state`,
              plus `iss` per RFC 9207.

            OAuth 2.1 §7.5.4 / RFC 9700 §4.11 RECOMMEND HTTP 303 over 302
            and PROHIBIT 307; both are modeled as alternative status codes.
          headers:
            Location:
              explode: false
              required: true
              schema:
                format: uri
                type: string
              style: simple
        "303":
          description: Same as 302; RECOMMENDED status code per OAuth 2.1 §7.5.4.
          headers:
            Location:
              explode: false
              required: true
              schema:
                format: uri
                type: string
              style: simple
        "200":
          content:
            text/html:
              schema:
                title: kid
                type: string
          description: HTML consent or authentication page (intermediate state)
        "400":
          content:
            text/html:
              schema:
                title: kid
                type: string
          description: |
            Non-redirecting error. Per RFC 6749 §4.1.2.1, when `redirect_uri`
            is missing, malformed, or does not match a registered URI, the
            AS MUST NOT redirect to it; instead it returns an HTML error page.
            Also returned for malformed `client_id`.
      summary: Authorization endpoint
      tags:
      - Core
  /token:
    post:
      description: |
        Issues access tokens. The `grant_type` form field selects the grant. All
        grants supported by OAuth 2.1 are modeled here:

        - `authorization_code` — RFC 6749 §4.1.3 (with PKCE per RFC 7636)
        - `client_credentials` — RFC 6749 §4.4
        - `refresh_token` — RFC 6749 §6
        - `urn:ietf:params:oauth:grant-type:device_code` — RFC 8628 §3.4
        - `urn:ietf:params:oauth:grant-type:jwt-bearer` — RFC 7523
        - `urn:ietf:params:oauth:grant-type:token-exchange` — RFC 8693

        The deprecated `password` and `implicit` grants are intentionally
        omitted (per OAuth 2.1).
      externalDocs:
        description: RFC 6749 §3.2
        url: https://datatracker.ietf.org/doc/html/rfc6749#section-3.2
      operationId: token
      parameters:
      - $ref: "#/components/parameters/DPoPHeader"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Access token issued
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
            DPoP-Nonce:
              $ref: "#/components/headers/DPoPNonce"
        "400":
          $ref: "#/components/responses/OAuthError"
        "401":
          $ref: "#/components/responses/OAuthError"
      security:
      - clientSecretBasic: []
      - clientSecretPost: []
      - clientAssertion: []
      - mtls: []
      - none: []
      summary: Token endpoint (all grants)
      tags:
      - Core
...
  /device_authorization:
    post:
      description: |
        Initiates the Device Authorization Grant. The client receives a
        `device_code` and `user_code`; the user is instructed to visit
        `verification_uri` on a second device and enter the `user_code`.
        The client then polls `/token` with
        `grant_type=urn:ietf:params:oauth:grant-type:device_code` until the
        user authorizes or `expires_in` elapses.

        Polling protocol (RFC 8628 §3.5): the AS may return `slow_down` to
        increase the polling interval by 5 seconds, or `authorization_pending`
        while the user is still consenting. Clients MUST stop polling on any
        non-pending error (`access_denied`, `expired_token`).

        The `device_code` MUST NOT be displayed to the end user.
        The `user_code` is per §6.1 displayed to the user; servers SHOULD
        accept dashed and case-insensitive variants.
      externalDocs:
        description: RFC 8628 §3.1
        url: https://datatracker.ietf.org/doc/html/rfc8628#section-3.1
      operationId: deviceAuthorization
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/DeviceAuthorizationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationResponse"
          description: Device authorization issued.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
        "400":
          $ref: "#/components/responses/OAuthError"
        "401":
          $ref: "#/components/responses/OAuthError"
      security:
      - clientSecretBasic: []
      - clientSecretPost: []
      - clientAssertion: []
      - mtls: []
      - none: []
      summary: Device Authorization Endpoint (RFC 8628)
      tags:
      - Device Flow
  /par:
    post:
      description: |
        Pushes the authorization request parameters to the server before the
        user-agent redirect. The server returns a short-lived `request_uri`
        that the client then passes (instead of the full request) to
        `/authorize`. This avoids leaking parameters via browser history,
        keeps URLs short, and enables integrity protection.

        Per §2.1, the request body MUST NOT contain a `request_uri` parameter.
        Per §4, when `/authorize` is reached via a PAR-issued `request_uri`,
        only `client_id` and `request_uri` are sent — all other parameters
        come from the pushed body.
      externalDocs:
        description: RFC 9126 §2
        url: https://datatracker.ietf.org/doc/html/rfc9126#section-2
      operationId: pushedAuthorizationRequest
      parameters:
      - $ref: "#/components/parameters/DPoPHeader"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PARRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PARResponse"
          description: Pushed authorization request accepted.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
            DPoP-Nonce:
              $ref: "#/components/headers/DPoPNonce"
        "400":
          $ref: "#/components/responses/OAuthError"
        "401":
          $ref: "#/components/responses/OAuthError"
        "405":
          description: Method Not Allowed (RFC 9126 §2.3) — only POST is permitted.
        "413":
          description: Payload Too Large (RFC 9126 §2.3) — request exceeds AS-defined
            limit.
        "429":
          description: Too Many Requests (RFC 9126 §2.3) — rate limit exceeded.
      security:
      - clientSecretBasic: []
      - clientSecretPost: []
      - clientAssertion: []
      - mtls: []
      - none: []
      summary: Pushed Authorization Request (RFC 9126)
      tags:
      - PAR
...
components:
...
  securitySchemes:
    bearerAuth:
      description: |
        OAuth 2.0 Bearer Token presentation per RFC 6750 §2.1
        (`Authorization: Bearer <token>` header). The token format is
        implementation-defined: it MAY be opaque (validated via
        introspection — RFC 7662) or a JWT (RFC 9068 §2 — `typ: at+jwt`,
        validated locally against `jwks_uri`). `bearerFormat` is intentionally
        omitted because both forms are common in the wild.
      scheme: bearer
      type: http
    dpop:
      description: |
        DPoP-bound access token (RFC 9449). The HTTP Authorization header
        carries `DPoP <access-token>`, and a SECOND `DPoP` header MUST also
        be sent containing a fresh proof JWT (typ=`dpop+jwt`). OpenAPI 3.1
        cannot natively model the second-header requirement under a single
        `securityScheme`; operations using DPoP also accept the `DPoP`
        request-header parameter (`components.parameters.DPoPHeader`) to
        document it. RFC 9449 §7.1 also defines a DPoP `WWW-Authenticate`
        challenge with an `algs` parameter for resource servers.
      in: header
      name: Authorization
      type: apiKey
    clientSecretBasic:
      description: |
        Client authentication via HTTP Basic with `client_id:client_secret`
        per RFC 6749 §2.3.1. PREFERRED over `clientSecretPost` per the same
        section.
      scheme: basic
      type: http
    clientSecretPost:
      description: |
        Client authentication via `client_id` + `client_secret` form fields
        in the request body (RFC 6749 §2.3.1). NOT RECOMMENDED — prefer
        `clientSecretBasic`. This scheme is a documentation marker only;
        OpenAPI lacks a body-credentials security type. In practice the
        fields appear in `ClientAuthFields` and the operation's request body.
      in: header
      name: X-OAuth-Client-Secret-Post
      type: apiKey
...